pyproject.toml
[tool.poetry]
name = "pytaku"
version = "0.1.3"
description = ""
authors = ["Bùi Thành Nhân <hi@imnhan.com>"]
license = "AGPL-3.0-only"
packages = [
    { include = "pytaku", from = "src" },
    { include = "mangoapi", from = "src" },
]

[tool.poetry.scripts]
pytaku-migrate = "pytaku:migrate"
pytaku-generate-config = "pytaku:generate_config"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
gunicorn = "^20.0.4"
requests = "^2.24.0"
goodconf = "^1.0.0"
argon2-cffi = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
